Final Adjustment GCodeProcessor
- I got confused when renaming the files
- this is the correct version now

Changes:

Renamed Gcode_class to GCodeProcessor.
The class has been updated based on the feedback from @liquidcronos, except for the Euler conversion and the methods for moving the end effector and robot.
It would be beneficial to have a meeting to discuss potential solutions for running those commands outside of the GCodeProcessor.